add: bin/rename_image_hashes.php you should run this immediately after updating else all of your images will be redownloaded and reprocessed fix: torrent name breaking template update: check if image is in person.photo before removing remove_altered_images.php add: cli options to delete, validate or optimize images validate_images.php add: images from person.photo optimize_resize_images.php update: change image proxy naming from sha512 to sha256 update: another attempt to fix FOUC update: check/updated to 1 week instead of 1 hour images_update.php fix: person.updated not updating get_imdb_person() add: rename images to sha256 bin/rename_image_hashes.php update: do not insert person photo when photo is nopicture update: run jobby log permission every 15 min instead of every min

+++ bin/validate_images.php
@@ -2,39 +2,68 @@
 
 declare(strict_types = 1);
 
+use Pu239\Database;
 use Pu239\ImageProxy;
 
 require_once __DIR__ . '/../include/bittorrent.php';
 global $container, $site_config;
 
+if (!isset($argv[1]) || ($argv[1] !== 'validate' && $argv[1] !== 'optimize' && $argv[1] !== 'purge')) {
+    die("This script can validate, optimize and delete all images found in public/images/proxy\n\nTo run:\n{$argv[0]} [purge|validate|optimze]\n\n");
+}
+foreach ($argv as $arg) {
+    $optimize = $arg === 'optimize' ? true : false;
+    $purge = $arg === 'purge' ? true : false;
+    $validate = $arg === 'validate' ? true : false;
+}
+if ($purge) {
+    $fluent = $container->get(Database::class);
+    $urls = $fluent->from('images')
+                   ->select('null')
+                   ->select('url');
+    $hashes = [];
+    foreach ($urls as $url) {
+        $hashes[] = hash('sha256', $url['url']);
+        if ($url['type'] === 'poster') {
+            $hashes[] = hash('sha256', $url['url'] . '_converted_' . '20');
+            $hashes[] = hash('sha256', $url['url'] . '_450');
+            $hashes[] = hash('sha256', $url['url'] . '_250');
+        } elseif ($url['type'] === 'banner') {
+            $hashes[] = hash('sha256', $url['url'] . '_1000');
+        }
+    }
+}
 $site_config['cache']['driver'] = 'memory';
 $image_proxy = $container->get(ImageProxy::class);
 $paths = [
-    IMAGES_DIR . 'proxy',
+    PROXY_IMAGES_DIR,
 ];
 
-$optimize = !empty($argv[1]) && $argv[1] === 'optimize' ? true : false;
-$dirsize = $o = $i = 0;
+$dirsize = $o = $i = $x = 0;
 foreach ($paths as $path) {
     $dirsize += GetDirectorySize($path, false, false);
     $objects = new RecursiveIteratorIterator(new RecursiveDirectoryIterator($path, RecursiveDirectoryIterator::SKIP_DOTS), RecursiveIteratorIterator::SELF_FIRST);
     foreach ($objects as $name => $object) {
-        if (!exif_imagetype($name)) {
-            if (basename($name) === '.gitignore') {
+        $basename = basename($name);
+        if ($purge && !in_array($basename, $hashes)) {
+            unlink($name);
+            ++$x;
+        }
+        if ($validate && !exif_imagetype($name)) {
+            if ($basename === '.gitignore') {
                 continue;
             }
             ++$i;
             unlink($name);
-            echo "$name \n";
-        } else {
-            if ($optimize) {
-                $image_proxy->optimize_image($name);
-            }
-            ++$o;
         }
+        if ($optimize) {
+            $image_proxy->optimize_image($name);
+        }
+        ++$o;
     }
 }
 $dirsize = mksize($dirsize);
 echo "$o images validated
 Images size: $dirsize
+$x images not found in db\n
 $i bad images removed\n";
